requirements.txt
discord.py
pymongo==3.12.0
Quart==0.15.1
Jinja2==3.0.1
dnspython==1.16.0
gunicorn==20.1.0
Hypercorn==0.11.2